+++ emperor/format.py
@@ -641,7 +641,7 @@ document.getElementById("logotable").style.display = 'none';
                 </tr>
                 <tr>
                     <td align="center">
-                        <button id="toggle-visibility-selection-button" onClick="toggleVisibleCategories()">Toggle Selected</button>
+                        <button id="toggle-visibility-selection-button" onClick="toggleVisibleCategories()">Invert Selected</button>
                         <br><br><br><br>
                     </td>
                 </tr>